Create a core RadixSortLSD that can sort arbitrary data
Previously, the main differences between `RadixSortLSD_keys` and
`RadixSortLSD_ranks` were transformations on the input data and how the
key was extracted from that transformed data. For key sorting no
transformations are done and the array element is just the key. For rank
sorting a tuple of key/rank is created and the key has to extracted.

This makes the core sort agnostic to those two aspects. Now the array
transformation is done before calling sort and a comparator is used to
extract the key. Comparators are based off the ones in Chapel's sort
https://chapel-lang.org/docs/modules/packages/Sort.html#comparators

Note that this core-sort is in-place so for cases where a transformation
is not needed, a copy must be made to avoid changing the input array.