Add a flag to get_datasets and ls_hdf for Parquet files This PR adds an `is_parquet` flag to both the `get_datasets()` function and the `ls_hdf()` function. This is not ideal since `ls_hdf` implies HDF5 files, but in the short term before we have figured out auto-detecting of Parquet files, I think that this approach makes sense.
In the C++/C side, there was a function added to return a
string of the dataset names separated by a comma of all
datasets that we support the types of today (int, uint,
timestamp). In the Chapel side, there was a `lspqMsg`
added that is very similar to the `lshdfMsg` function,
but with the Parquet calls substituted in. This means
that once we have the auto detecting of Parquet files
figured out, it will be easy to combine the two functions.