Eliminate a copy for the scan on globalStarts in sort
Today scan always creates a new array and returns it. By assigning the
result of a scan into an existing array we were just adding an array
copy and not eliminating an allocation. Switch to just assigning into a
new variable, which will steal the returned buffer and avoid the init
allocation for globalStarts and the assignment to it during scans.

Longer term, we have discussed having a way to provide the output array
for scans, but there's no timeframe for that.