SegArray moved into Arkouda proper
Removed register/attach functionality because not supported for GroupBy. Added comment to indicate the status.

Adding dataframe class to arkouda proper. Remove GroupBy class that appeared to be legacy code. All other code is using the arkouda groupby and no refereneces made to the class found in akutil/dataframe.py.