Make some minor optimizations to RadixSortLSD (#1120)
* Avoid an extra copy for sort

Part of 1099 added an extra copy between the input array and the temp
array. Remove that here by switching the order the arrays are used so we
end up PUT'ing sorted results into the input/return array instead of the
temp one.

* Eliminate a copy for the scan on globalStarts in sort

Today scan always creates a new array and returns it. By assigning the
result of a scan into an existing array we were just adding an array
copy and not eliminating an allocation. Switch to just assigning into a
new variable, which will steal the returned buffer and avoid the init
allocation for globalStarts and the assignment to it during scans.

Longer term, we have discussed having a way to provide the output array
for scans, but there's no timeframe for that.

* Use localAccess for some array accesses in sort

`localAccess` is a minor optimization that avoids some locality checks
when array accesses are local. The performance benefit is small, but not
trivial and given the importance of sort it seemed worthwhile.