Tune int in1d threshold
Tune the threshold for selecting in1d implementations. Even with recent
optimizations, the threshold for selecting between the set and sort
based strategies is a little too high. Drop from `2**25` down to
`2**23`, which is pretty good on 16-node-cs-hdr, 16-node-xc, and
512-node-xc. The ideal threshold depends on core count, node count, and
network bandwidth, but I think this will work well in practice. For some
configs `2**24` may be a little better, but we I'd prefer to stick lower
since we know the sort based methods scales well, it just has high
startup overhead.

Part of 970