Add support for reading string columns from Parquet files This PR adds the initial, unoptimized code for reading string columns from Parquet files using the same `ak.read_parquet()` interface as the other Parquet dtypes.
For this initial implementation, performance was not a focus
in an attempt to add functionality and iteratively improve
performance, but it is worth noting that the performance and
scalability of this code is not good with this initial support
PR. The `segmentedCalcOffsets()` function (that is used in HDF5
as well) is used and it appears to have a large amount of comm
and does not scale as locales are added. Additionally, there does
not seem to be a query to figure out the byte size of a Parquet
string column, so we have to pass over the file twice; first to
get the number of bytes per row to be able to construct the values
array prior to reading in the values, then, a second time, to
actually read in the data.

Since there is no support for writing Parquet string columns in
Arkouda today, I wasn't able to come up with a good Arkouda unit
test for this functionality, but there is a Chapel unit test
testing the C++ code.