Fix crash when passing in bad column names to Parquet (#1157)
When calling `ak.read_parquet()` with a bad column name
that was not the first column was causing the server to
crash. This is because only the first dataset name in the
set of dataset names was being checked, so if dataset
names after the first were bad, they would cause the
server to crash.

This is a remnant of copy-pasting code from the HDF5
layer that I did not totally understand that does not
work identically for Parquet.

Now, each dataset name will be checked and an error
will be returned if the dataset does not exist in the
file, rather than crashing the server.

Additionally, I have removed a bunch of the error handling
code that is HDF5 specific, as none of the Parquet code
throws anything other than a basic `Error`.