Remove trailing space for ak.get_datasets (#1159)
When implemented `ak.get_datasets` for Parquet,
I had left a trailing space in the string of
dataset names that was constructed in the C++ code,
which was causing the column names to incorrect for
all but the first column.

This slipped through testing because there is currently
no way to write multi-column Parquet files in Arkouda
and there is no support for appending to Parquet files
today. If we were able to add that functionality, testing
for Parquet would become much stronger.