Add append mode for Parquet file writing (#1227)
This PR adds append mode for all dtypes in Parquet besides
strings. Since Parquet does not support any method of appending
columns to a file, all values are read from the file before
appending the column and rewriting out the entire file. Due to
this, as well as using the higher level `ReadTable` and
`WriteTable` functions instead of `ReadBatch` and `WriteBatch`,
the performance of writing Parquet files using append mode is
significantly worse than writing a completely new Parquet file.

Some additional limitations with the Parquet implementation
require that columns appended to an existing match the number
of rows of other columns and that the dataset name is not
duplicated in the file (in other words, this only allows you
to add a new column to the file, not add values to an existing
column).

command                     | exec time 1 node | exec time 8 nodes |
--------------------------- | ----------------:| -----------------:|
truncate (default)          | 0.0967 s         | 0.072 s           |
append with 1 existing col  | 0.42 s	       | 0.106 s           |
append with 2 existing cols | 0.59 s           | 0.110 s           |