Fix append Parquet test when running with more than 10 locales (#1237)
The Parquet unit test was only writing with 10 elements, which
meant that when running with greater than 10 nodes, there would
be some files that weren't created or were only created with
string files. This PR adds an error when trying to append an
empty column with Parquet strings and also bumps the unit test
size so it works with greater node counts.