Unify various file reading functions into a single read function (#1288)
* Unify various file reading functions into a single read function
This PR takes `read_hdf()`, `read_all()`, and `read_parquet()` and
combines them into a single `read()` function. This new function is
identical to the old `read_all()` function with the exception of the
`file_format` keyword argument. This defaults to `"infer"`, which
means that the server will determine the type of the file and it does
not need to be specified by the user.

This file type inference changes certain error checking behaviors
(e.g., if you have an error in one of your filenames, this will behave
differently that before with `allow_errors=True`). To still allow the
old functionality, the `file_format` keyword argument can be explicitly
set to either `HDF5` or `Parquet` (case insensitive), and it will
behave as before.

Additionally, an `ak.get_filetype()` function has been added that will
return the type of a file, or the type of the first file in a glob
expression.

* Fix mypy issues

* Update benchmark with new read calls

* Fix some last instances of read_all

* Fix python 3.7 compatibility issue

* Remove double read from docs

* Update util.py

Co-authored-by: Michael Merrill <48804335+mhmerrill@users.noreply.github.com>