Update strings object to match new save function convention (#1307)
In #1293, the `pdarray.save()` method was updated with a `file_format`
keyword that allowed specification of Parquet or HDF5 files to make
the API simpler to use. In that PR, strings were overlooked and had
not been updated to match this new API design.

This PR updates the `strings.save()` function to use a `file_format`
keyword argument like the other PR and also adds `save_hdf` and
`save_parquet` wrappers around that function.

Additionally documentation has been updated and a discrepency between
the HDF5 strings/pdarray `tohdf` command arguments was cleaned up. An
unused argument was removed and the strings and pdarray message
arguments now send the same number of arguments (7) to the server,
whereas they used to send differing numbers (6 for pdarrays and 8
for strings).