Closes #1393: Implement assume_sorted shortcut for groupby and unique
This PR (Closes #1393):
- Adds `assume_sorted` flag to `unique` and passes this through `GroupBy.__init__`
- Adds `assumeSortedShortcut` which behaves like `uniqueAndCount` but without calling radixSort
- Adds memcheck for `uniqueAndCount`

Note:
It's very likely this is not the most efficient way to skip sorting. I wanted to go ahead and get this up to fix the out of memory errors seen by users. I would love feedback on a better way to implement this especially from reuster986