Limit multi-locale CI testing
The gasnet correctness testing is now taking over an hour. Back in
January this was down to 30 minutes, which felt reasonable but an hour
is too much and limits our ability to get new code in. For now, limit
multi-locale testing to `make check` and rely on nightly testing to run
the full test suite.

I didn't want to remove all multi-locale testing since I think there's
value in some checks with the different distributions and aggregation
being used, but the full test suite just takes too long at the moment.

Part of 388
Closes 1417