Update to fix test/UnitTestSort.chpl with Chapel main
That test was complaining it could not find AryUtil, because the 1.27
pre-release has changed our handling of `public use` to no longer bring in the
name of the module in addition to the symbols it contains, unless an `as` clause
is used.  Add an `as` clause so that the name is now brought in again.

I believe this was missed due to not running testing locally.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>