Adjust source files to be compatible with changes I made to the Math module (#1438)
* Adjust source files to be compatible with changes I made to the Math module

The Chapel team is in the process of splitting the Math standard library into a
portion that continues to be included by default and one that will require a
`use` or `import` statement going forward.  To maintain compatibility with both
the previous releases and main/the upcoming release, update these files to no
longer use explicit naming to access the symbols in that module (since the ones
that are being included by default are no longer in the module named "Math" but
are instead in a new module named "AutoMath" that won't exist for previous
releases)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>

* Update to fix test/UnitTestSort.chpl with Chapel main

That test was complaining it could not find AryUtil, because the 1.27
pre-release has changed our handling of `public use` to no longer bring in the
name of the module in addition to the symbols it contains, unless an `as` clause
is used.  Add an `as` clause so that the name is now brought in again.

I believe this was missed due to not running testing locally.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>