Use swap operator in radix sort instead of assignment operator (#1419)
This PR switches an assignment operator to use the swap operator
instead. The swap operator is a simple pointer swap, whereas the
assignment operator is a deep copy, so nearly all of the time
from that assignment is saved, but it is admittedly not the
bottleneck, so only modest performance improvements should be
expected.