Update ancillary files to avoid using SysCTypes, CPtr
This is a follow-up on PR #1185 which updated Arkouda to use with
the new 'CTypes' module introduced in Chapel 1.26.  Specifically,
I'd overlooked that other files still relied on SysCTypes and CPtr,
such as those used for the check-* rules in the Makefile.

Here, I'm updating those files to also use the new 'CTypes' modules
and updating the Makefile to add the ARKOUDA_COMPAT_MODULES module
search path to their compilations as well.  This seems like a
reasonable change in that (as I understand it) the point of the
check-* rules is to see whether the main compilation will work, so
making the 'chpl' command lines more similar seems like it's only
making the check that much better (?).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>