Always recompile Arrow when running make or make check-arrow (#1521)
This PR makes it so that the Arrow C++ code will always recompile
on a `make` or `make check-arrow`. When upgrading Arrow versions,
if the C++ code is not recompiled, you could end up having problems
with the version that the C++ code was compiled with and the version
that the Chapel code is trying to link with, which will cause
problems.

Also, this PR adds removal of `ArrowFunctions.o` to `make clean`.