Switch to using a lowLevelLocalizingSlice for DF indexing message (#1537)
This PR is a performance improvement for the `_get_head_tail_server`
function. Since we know that the index arrays will be small that
we are using to index in the DataFrameIndexingMsg function, we can
hardcode in that expectation to avoid some overhead that is helpful
only when doing large-array indexing.

To get around this overhead, a `smallIdx` argument was added to
pdarray string indexing by integer arrays indicating that the
indexing array is small and would not benefit from all of the
large array index overhead.

Instead of doing the regular indexing, the `smallIdx` branch will
localize each string using `lowLevelLocalizingSlice` before copying
it into the output array. This performs better than the naive
indexing approach taken when `CHPL_COMM=none` for distributed runs
for arrays over a certain threshold and significantly outperforms
the large array path.

This optimization could be extended to dynamically check the size of
the index array and then determine whether or not we want to do the
indexing optimized for large or small arrays, but pinning down those
thresholds can be difficult and machine-dependent.

Running 50 `_get_head_tail_server` string calls on 8
nodes of a cs-hdr:
String size | This branch | Naive copy | Master
----------- | ----------- | ---------- | ------
5-6         | 0.7960 s    | 0.7883 s   | 1.1169 s
50-51       | 0.8262 s    | 1.0042 s   | 1.2797 s
100-101     | 0.8736 s    | 1.2195 s   | 1.3165 s
200-201     | 0.8778 s    | 1.6184 s   | 1.3511 s
300-301     | 0.9518 s    | 1.9948 s   | 1.4371 s
1000-1001   | 1.363 s     | --         | 1.8706 s
2000-2001   | 1.8799 s    | --         | 2.3605 s

Running `dataframe.py` benchmark on 16 nodes of a cs-hdr:
This branch | Master
----------- | ---------
0.8991 s    | 1.0048 s