Recommend Chapel 1.26.0 and use it for CI testing
Chapel 1.26 was mostly about new GPU features and better `CHPL_COMM=ofi`
support, neither of which are important to Arkouda currently, which is
largely why I haven't upgraded until now.

Note that this changes the documentation to "require" 1.26.0, but only
recommends it in the Makefile instead of making it an error. Arkouda
should continue to work with Chapel 1.24+ so existing users don't
necessarily have to upgrade, but I think it makes sense to recommend
that users reading the doc should use the latest version.

Note that 1.24 is pretty old at this point and we can probably drop some
support for some older versions, but I'll leave that for others.

Resolves 1543