Quiet deprecation warnings for Chapel 1.27 (#1553)
* Change `datetime.today()` to `datetime.now()`

`datetime.today()` is deprecated in Chapel 1.27 because it has identical
behavior to the existing `datetime.now()` and we decided to prefer that
name. Change to it to avoid the deprecation warning and future removal.

* Change `string.match(regex)` to `regex.match(string)`

In Chapel 1.27 `string.match(regex)` was deprecated in favor of
`string.startsWith(regex)`. However, there is also `regex.match(string)`
which has the same behavior and has existed for longer, so switch to
that to avoid having to add a compatibility shim for `startsWith`

* Avoid calling .low/.high on strided ranges/domains

The definition of low/high on strided ranges and domains will change to
the aligned low/high bound instead of the pure bound. For example the
low bound of `1..10 by -2` will be 2 instead of 1. Chapel 1.27 adds a
warning about this upcoming change, so adjust Arkouda to avoid calling
low/high on strided ranges to quiet the warning.

The changes in AryUtil are trivial and the `stridable=false` change
isn't necessary for the warnings, but it occurred to me the code only
works for non-strided ranges so I wanted to document this. For string
slices, I noticed that the code errors if the stride from python is
anything other than 1 so we don't actually need the chapel code to be
strided today. If the support for strided string slices is ever added in
the future, the range can be made strided again (but hopefully when that
functionality is needed these warnings will be gone and we can just use
low/high or whatever is appropriate then.)