Rename `SegString.hash()` to `SegString.siphash()`
`hash()` is currently a special method in Chapel and we recently added
some extra checking upstream that disallows using the name without the
`override` keyword. We haven't settled on the name `hash()` yet and are
considering special methods with names more unique as to not interfere
with user names, but until that's settled I think it's easiest to just
give the segmented string hash a different name to stay compatible.

Resolves 1555