Closes #1548: Fix `interval_lookup` bug and enable multi-array input (#1549)
This PR (Closes #1548):
- Updates calls to `search_intervals` to not use `assume_unique` (this keyword arg was removed in #1494)
- Adds ability to pass multi-array input into `interval_lookup` and `search_intervals`
- Adds a test for the multi-array input

Note:
I'll admit this implementation is definitely not the most elegant. Since this is an important bug, I went with what seemed most straightforward to me, but I fully expect there's a lot that can be done to make this code more efficient and prettier. Stylisitic suggestions are very welcome

Co-authored-by: Pierce Hayes <pierce314159@users.noreply.github.com>