Add portability testing to older chpl versions (#1562)
Add portability testing for chpl 1.25/1.26. The documentation currently
"requires" 1.27, but we expect older versions to continue to work. Add
some testing to lock in that the versions we expect to work continue to
do so. I think we're planning to drop support for 1.25 soon, but wanted
to start with it for now.

Closes 1545