Closes #1519: Add strip method to Strings
This PR (Closes #1519):

- Add ability to strip leading and trailing characters from original Strings
        - If a pattern is provided, it is used as a list of leading and trailing characters
        - If no pattern is provided, the leading and trailing characters default to spaces

Note:
The given pattern can include any whitespace characters. The space character (" ") is sent
to chapel in the form of a flag indicating the space character is included with the pattern
since arkoulda does not pass spaces embedded in arguments when passing messages from the
client to server.