Remove python script generation for Chapel unit tests (#1575)
`CommandMap.chpl` for some reason had a `use ServerRegistration`
that I believe was added by accident and that was making the
Chapel unit tests need the `ServerRegistration` script to be
generated. This PR removes that unnecessary `use` and takes out
the python script call from the CI and Chapel unit tests.