Closes #1572: Update modular build process to initialize modules on command line (#1606)
* Update modular build process to initialize modules on command line
This PR changes from the current approach of generating a
`ServerRegistration.chpl` file that calls the `registerMe()`
function required in all modules to instead move the contents
of the `registerMe()` to module-level scope that will add the
functions to the command map upon module initialization. All
modules that are to be modularly included in the server are now
directly specified on the command line.

This change is enabled by the work from https://github.com/chapel-lang/chapel/pull/18724,
which initializes all modules specified on the command line of a
`chpl` call, even if the modules aren't explicitly `use`d. This
means that Arkouda will require at least Chapel 1.26 after this
change.

* Convert rest of modules

* Update python script to handle non-src dir files

* Remove 1.25.1 compat check

* Add space caught by Pierce

* Update modular building docs

* Update MODULAR.md

fixed a typo `pat` -> `path`

Co-authored-by: pierce314159 <48131946+pierce314159@users.noreply.github.com>