Part of 1901: Replace `SymEntry.Ad` with `SymEntry.a.domain`
Motivated by 1901, I'd like to stop storing the domain in the `SymEntry`
wrapper, so we now need to go through the array to get the domain (e.g.
`symEntry.aD` is now `symEntry.a.domain`). You could imagine creating a
wrapper to do this automatically with a wrapper like:

```chpl
proc aD const ref {
  return a.domain;
}
```

But unfortunately chapel-lang/chapel 20389 prevents using `const ref`,
which is required to avoid a full domain copy on each call.

Part of 1901