Closes #1781 - Quiet deprecation warnings in prep for Chapel 1.28 (#1782)
* Handle DateTime deprecation

The `DateTime` module has been deprecated in favor of `Time`. Extend the
existing `ArkoudaDateTimeCompat` module to paper over the transition
period.

* Handle SysError deprecation

The `SysError` module has been deprecated, but it turns out Arkouda
didn't actually need it anymore, so remove the remaining `use`s.

* Handle isFloatType deprecation

The `isFloatType` function is deprecated in favor of a more explict
`isRealType || isImagType` query, but Arkouda only supports `real`s
anyways, so update the checks to just look for real's.

* Handle owned->shared class assignment deprecation

Assignment from owned to shared has been deprecated, so update Arkouda's
use to explicitly create shared objects in the first place instead of
getting it through assignment (implicit in the function return)

* Stop using `start` for file.reader

`file.reader` has changed from explicit ints for `start`/`stop` to a
range to describe the `region`. The few uses in arkouda were specifying
a start of 0, which is the default so we can just stop specifying it use
the default.

* Handle file.writer start -> region change

`file.writer` has changed from explicit ints for `start`/`stop` to a
range to describe the `region`. Use explicit version checks to select
which version to use in order to compatible across versions.

* Stop relying on `<~>` operator

The `<~>` operator has been deprecated, so stop using it an just rely on
explicit `write` calls.

* Avoid using the deprecated `list.extend`

`list.extend` was deprecated in favor of a `list.append` overload, but
that requires at least chapel 1.27, so to maintain compatibility to 1.26
just explicitly loop over and do individual `append` calls.