Closes #1837: Intermittent failures in groupby prod aggregate test
This PR (Closes #1837)
- Updates code using `equality_helper` to use `np.allclose(equal_nan=True)` instead

A github workflow in my fork saw a failure in the groupby prod aggregation in `test_groupby_on_one_level`
```
Doing aggregate(uint64, prod)
Different values (abs diff = 1.8446744073709564e+19)
```
I'm not sure if I'm just misinterpretting, but I think this should be within the tolerance of `np.allclose`. So I'm not sure how this happened. While looking into this, I noticed the `equality_helper` is not necessary if we use the `equal_nan` flag. So I went ahead and updated this (and spots in operator_test) to use the functionality built into numpy instead